feat: adicionar policies, autorização nos controllers e cálculo de multa

+++ app/Http/Controllers/BrandController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreBrandRequest;
 use App\Http\Requests\UpdateBrandRequest;
 use App\Http\Resources\BrandResource;
+use App\Models\Brand;
 use App\Repositories\Contracts\BrandRepositoryInterface;
 use Illuminate\Http\Request;
 
@@ -19,6 +20,8 @@ class BrandController extends Controller
 
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Brand::class);
+
         if ($request->has('name')) {
             return BrandResource::collection($this->repository->search($request->name));
         }
@@ -28,22 +31,33 @@ class BrandController extends Controller
 
     public function store(StoreBrandRequest $request)
     {
+        $this->authorize('create', Brand::class);
+
         $brand = $this->repository->create($request->validated());
         return (new BrandResource($brand))->response()->setStatusCode(201);
     }
 
     public function show($id)
     {
-        return new BrandResource($this->repository->find($id));
+        $brand = $this->repository->find($id);
+        $this->authorize('view', $brand);
+
+        return new BrandResource($brand);
     }
 
     public function update(UpdateBrandRequest $request, $id)
     {
+        $brand = $this->repository->find($id);
+        $this->authorize('update', $brand);
+
         return new BrandResource($this->repository->update($id, $request->validated()));
     }
 
     public function destroy($id)
     {
+        $brand = $this->repository->find($id);
+        $this->authorize('delete', $brand);
+
         $this->repository->delete($id);
         return response()->json(['msg' => 'A marca foi removida com sucesso!'], 200);
     }
