+++ app/Http/Controllers/CarController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\ActiveRentalsException;
 use App\Http\Requests\StoreCarRequest;
 use App\Http\Requests\UpdateCarRequest;
 use App\Http\Resources\CarResource;
+use App\Models\Car;
 use App\Repositories\Contracts\CarRepositoryInterface;
 use Illuminate\Http\Request;
 
@@ -19,6 +21,8 @@ class CarController extends Controller
 
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Car::class);
+
         if ($request->has('plate')) {
             return CarResource::collection($this->repository->searchByPlate($request->plate));
         }
@@ -32,22 +36,37 @@ class CarController extends Controller
 
     public function store(StoreCarRequest $request)
     {
+        $this->authorize('create', Car::class);
+
         $car = $this->repository->create($request->validated());
         return (new CarResource($car))->response()->setStatusCode(201);
     }
 
     public function show($id)
     {
-        return new CarResource($this->repository->find($id));
+        $car = $this->repository->find($id);
+        $this->authorize('view', $car);
+
+        return new CarResource($car);
     }
 
     public function update(UpdateCarRequest $request, $id)
     {
+        $car = $this->repository->find($id);
+        $this->authorize('update', $car);
+
         return new CarResource($this->repository->update($id, $request->validated()));
     }
 
     public function destroy($id)
     {
+        $car = $this->repository->find($id);
+        $this->authorize('delete', $car);
+
+        if ($car->rentals()->whereNull('period_actual_end_date')->exists()) {
+            throw new ActiveRentalsException('Não é possível remover um veículo com locação ativa.');
+        }
+
         $this->repository->delete($id);
         return response()->json(['msg' => 'O veículo foi removido com sucesso!'], 200);
     }
