+++ app/Http/Controllers/ClientController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\ActiveRentalsException;
 use App\Http\Requests\StoreClientRequest;
 use App\Http\Requests\UpdateClientRequest;
 use App\Http\Resources\ClientResource;
+use App\Models\Client;
 use App\Repositories\Contracts\ClientRepositoryInterface;
 use Illuminate\Http\Request;
 
@@ -19,6 +21,8 @@ class ClientController extends Controller
 
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Client::class);
+
         if ($request->has('name')) {
             return ClientResource::collection($this->repository->search($request->name));
         }
@@ -28,22 +32,37 @@ class ClientController extends Controller
 
     public function store(StoreClientRequest $request)
     {
+        $this->authorize('create', Client::class);
+
         $client = $this->repository->create($request->validated());
         return (new ClientResource($client))->response()->setStatusCode(201);
     }
 
     public function show($id)
     {
-        return new ClientResource($this->repository->find($id));
+        $client = $this->repository->find($id);
+        $this->authorize('view', $client);
+
+        return new ClientResource($client);
     }
 
     public function update(UpdateClientRequest $request, $id)
     {
+        $client = $this->repository->find($id);
+        $this->authorize('update', $client);
+
         return new ClientResource($this->repository->update($id, $request->validated()));
     }
 
     public function destroy($id)
     {
+        $client = $this->repository->find($id);
+        $this->authorize('delete', $client);
+
+        if ($client->rentals()->whereNull('period_actual_end_date')->exists()) {
+            throw new ActiveRentalsException('Não é possível remover um cliente com locação ativa.');
+        }
+
         $this->repository->delete($id);
         return response()->json(['msg' => 'O cliente foi removido com sucesso!'], 200);
     }
